fix: type PolicyLeaveUsageDto.durationDays as Float to match the entity
The projection declared Double while lv_leave_request.duration_days maps to Float,
which needed a .as(Double.class) hint on every usage query to find a matching
constructor. Typing the field to match the column drops both casts and the
floatValue conversion in sumUsage.

+++ backend/src/main/java/com/skapp/community/leaveplanner/service/impl/PolicyLeaveServiceImpl.java
@@ -860,7 +860,7 @@ public class PolicyLeaveServiceImpl implements PolicyLeaveService {
 		for (PolicyLeaveUsageDto usage : usages) {
 			if (usage.getDurationDays() != null && !usage.getStartDate().isBefore(from)
 					&& !usage.getStartDate().isAfter(to)) {
-				totalDaysUsed += usage.getDurationDays().floatValue();
+				totalDaysUsed += usage.getDurationDays();
 			}
 		}
 		return totalDaysUsed;
